requirements.txt
streamlit
nltk
scikit-learn
networkx
gensim
numpy
PyPDF2
python-pptx